Use None as outputFile value for STDOUT
The comment "Use empty string rather then None to ease output to log."
is somehow outdated since Python 2.7+ supports None in '%s' formatting
and '{}' formatting

python2 -c 'print("Tracing: %s" % None)'
python2 -c 'print("Tracing: {}".format(None))'
python3 -c 'print("Tracing: %s" % None)'
python3 -c 'print("Tracing: {}".format(None))'

Just string concatenation does not work:

python2 -c 'print("Tracing: " + None)'
python3 -c 'print("Tracing: " + None)'

But we don't need this one anyway.

+++ clcache.py
@@ -280,7 +280,7 @@ class ObjectCache(object):
 
     def setEntry(self, key, objectFileName, compilerOutput, compilerStderr):
         ensureDirectoryExists(self._cacheEntryDir(key))
-        if objectFileName != '':
+        if objectFileName is not None:
             copyOrLink(objectFileName, self.cachedObjectName(key))
         with open(self._cachedCompilerOutputName(key), 'wb') as f:
             f.write(compilerOutput.encode(CACHE_COMPILER_OUTPUT_STORAGE_CODEC))
@@ -825,9 +825,8 @@ class CommandLineAnalyzer(object):
                 else:
                     outputFile = basenameWithoutExtension(sourceFiles[0]) + ".i"
             else:
-                # Preprocess to stdout. Use empty string rather then None to ease
-                # output to log.
-                outputFile = ''
+                # Preprocess to stdout
+                outputFile = None
         else:
             if 'Fo' in options:
                 outputFile = options['Fo'][0]
@@ -838,7 +837,7 @@ class CommandLineAnalyzer(object):
             else:
                 outputFile = basenameWithoutExtension(sourceFiles[0]) + ".obj"
 
-        printTraceStatement("Compiler output file: '%s'" % outputFile)
+        printTraceStatement("Compiler output file: {}".format(outputFile))
         return AnalysisResult.Ok, sourceFiles[0], outputFile
 
 
@@ -1063,10 +1062,9 @@ def parseIncludesList(compilerOutput, sourceFile, baseDir, strip):
 
 
 def addObjectToCache(stats, cache, outputFile, compilerStdout, compilerStderr, cachekey):
-    printTraceStatement("Adding file " + outputFile + " to cache using " +
-                        "key " + cachekey)
+    printTraceStatement("Adding file {} to cache using key {}".format(outputFile, cachekey))
     cache.setEntry(cachekey, outputFile, compilerStdout, compilerStderr)
-    if outputFile != '':
+    if outputFile is not None:
         stats.registerCacheEntry(os.path.getsize(outputFile))
         cfg = Configuration(cache)
         cache.clean(stats, cfg.maximumCacheSize())
@@ -1076,9 +1074,8 @@ def processCacheHit(cache, outputFile, cachekey):
     stats = CacheStatistics(cache)
     stats.registerCacheHit()
     stats.save()
-    printTraceStatement("Reusing cached object for key " + cachekey + " for " +
-                        "output file " + outputFile)
-    if outputFile != '':
+    printTraceStatement("Reusing cached object for key {} for output file {}".format(cachekey, outputFile))
+    if outputFile is not None:
         if os.path.exists(outputFile):
             os.remove(outputFile)
         copyOrLink(cache.cachedObjectName(cachekey), outputFile)
@@ -1089,14 +1086,13 @@ def processCacheHit(cache, outputFile, cachekey):
 
 
 def postprocessObjectEvicted(cache, outputFile, cachekey, compilerResult):
-    printTraceStatement("Cached object already evicted for key " + cachekey + " for " +
-                        "output file " + outputFile)
+    printTraceStatement("Cached object already evicted for key {} for output file {}".format(cachekey, outputFile))
     returnCode, compilerOutput, compilerStderr = compilerResult
 
     with cache.lock:
         stats = CacheStatistics(cache)
         stats.registerEvictedMiss()
-        if returnCode == 0 and (outputFile == '' or os.path.exists(outputFile)):
+        if returnCode == 0 and (outputFile is None or os.path.exists(outputFile)):
             addObjectToCache(stats, cache, outputFile, compilerOutput, compilerStderr, cachekey)
         stats.save()
 
@@ -1108,7 +1104,7 @@ def postprocessHeaderChangedMiss(cache, outputFile, manifest, manifestHash, keyI
     returnCode, compilerOutput, compilerStderr = compilerResult
 
     removedItems = []
-    if returnCode == 0 and (outputFile == '' or os.path.exists(outputFile)):
+    if returnCode == 0 and (outputFile is None or os.path.exists(outputFile)):
         while len(manifest.hashes) >= MAX_MANIFEST_HASHES:
             _, objectHash = manifest.hashes.popitem()
             removedItems.append(objectHash)
@@ -1117,7 +1113,7 @@ def postprocessHeaderChangedMiss(cache, outputFile, manifest, manifestHash, keyI
     with cache.lock:
         stats = CacheStatistics(cache)
         stats.registerHeaderChangedMiss()
-        if returnCode == 0 and (outputFile == '' or os.path.exists(outputFile)):
+        if returnCode == 0 and (outputFile is None or os.path.exists(outputFile)):
             addObjectToCache(stats, cache, outputFile, compilerOutput, compilerStderr, cachekey)
             cache.removeObjects(stats, removedItems)
             cache.setManifest(manifestHash, manifest)
@@ -1143,7 +1139,7 @@ def postprocessNoManifestMiss(
     manifest = None
     cachekey = None
 
-    if returnCode == 0 and (outputFile == '' or os.path.exists(outputFile)):
+    if returnCode == 0 and (outputFile is None or os.path.exists(outputFile)):
         # Store compile output and manifest
         manifest = Manifest(listOfIncludes, {})
         listOfHeaderHashes = [getRelFileHash(fileName, baseDir) for fileName in listOfIncludes]
@@ -1154,7 +1150,7 @@ def postprocessNoManifestMiss(
     with cache.lock:
         stats = CacheStatistics(cache)
         stats.registerSourceChangedMiss()
-        if returnCode == 0 and (outputFile == '' or os.path.exists(outputFile)):
+        if returnCode == 0 and (outputFile is None or os.path.exists(outputFile)):
             # Store compile output and manifest
             addObjectToCache(stats, cache, outputFile, compilerOutput, compilerStderr, cachekey)
             cache.setManifest(manifestHash, manifest)
@@ -1322,9 +1318,8 @@ def processNoDirect(cache, outputFile, compiler, cmdLine):
             stats = CacheStatistics(cache)
             stats.registerCacheHit()
             stats.save()
-            printTraceStatement("Reusing cached object for key " + cachekey + " for " +
-                                "output file " + outputFile)
-            if os.path.exists(outputFile):
+            printTraceStatement("Reusing cached object for key {} for output file {}".format(cachekey, outputFile))
+            if outputFile is not None and os.path.exists(outputFile):
                 os.remove(outputFile)
             copyOrLink(cache.cachedObjectName(cachekey), outputFile)
             compilerStdout = cache.cachedCompilerOutput(cachekey)
@@ -1336,7 +1331,7 @@ def processNoDirect(cache, outputFile, compiler, cmdLine):
     with cache.lock:
         stats = CacheStatistics(cache)
         stats.registerCacheMiss()
-        if returnCode == 0 and os.path.exists(outputFile):
+        if returnCode == 0 and outputFile is not None and os.path.exists(outputFile):
             addObjectToCache(stats, cache, outputFile, compilerStdout, compilerStderr, cachekey)
         stats.save()
